refactor(TestFixture): 패키지, 클래스명 수정
