+++ src/test/java/roomescape/fixture/TestFixture.java
@@ -1,4 +1,4 @@
-package roomescape.constant;
+package roomescape.fixture;
 
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
@@ -7,7 +7,7 @@ import roomescape.domain.Theme;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public class TestConstants {
+public class TestFixture {
 
     public static final ReservationTime DEFAULT_TIME = new ReservationTime(LocalTime.of(13, 5));
     public static final Theme DEFAULT_THEME = new Theme("폭싹 속았수다", "감동적인 내용", "abc");
